+++ views/news/index.php
@@ -32,12 +32,12 @@ $this->params['breadcrumbs'][] = $this->title;
                 'format' => 'raw',
                 'value' => function($model) {
                     $output = Html::tag('strong', $model->name);
-                    if (($pageURL = $model->getPostUrl()) && $model->id)
-                        $output .= '<br/>' . Html::a($pageURL, $pageURL, [
+                    if (($postURL = $model->getPostUrl(true, true)) && $model->id) {
+                        $output .= '<br/>' . Html::a($model->getPostUrl(true, false), $postURL, [
                                 'target' => '_blank',
                                 'data-pjax' => 0
                             ]);
-
+                    }
                     return $output;
                 }
             ],