+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog
 =========
 
 ## 1.0.2 (2019-08-06)
+ * Added draft view from backend
  * Added news image and thumbnails
  
 ## 1.0.1 (2019-08-02)